requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
pypdf>=3.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
